chore(engine): rename existing engine implementation to "basic"
While we work on exposing the engine using the scheduler and workers, we
need a new name for the original form of the engine. This renames the
original form to the "basic" engine, which at this point no longer
contains any parallelism.

Over time, this implementation will be removed in favour of a new form
which always uses the scheduler and workers.

Signed-off-by: Robert Fratto <robertfratto@gmail.com>