chore(engine): fix race condition on task assignment
The worker may hit a race condition depending on what the scheduler is
doing upon task creation:

* The worker marks itself as ready, holding on to `reqsMut` until the
  request is ackwnoledged.

* The worker is trying to process task assignment, but is blocked
  waiting to grab the lock on `reqsMut`.

The combination of these two conditions caused requests to time out,
since the scheduler couldn't process the worker being ready until the
task was assigned, but the task couldn't be assigned until the worker
received a ready message.

We fix this by not holding onto reqsMut for the entirety of marking a
worker as ready. This does mean that a different readyRequest may be
rejected if a ready request isn't acknowledged, but since a readyRequest
is not necessarily tied to a specific event, this is acceptable; threads
will retry like normal.

Signed-off-by: Robert Fratto <robertfratto@gmail.com>